start task and add observer
