+++ Src/Core/event_type.py
@@ -1,0 +1,12 @@
+from enum import Enum
+from Src.Core.validator import validator
+
+
+"""
+Типы событий
+"""
+class event_type(Enum):
+    DELETE_NOMENCLATURE = 1
+    CHANGE_NOMENCLATURE = 2
+    CHANGE_RANGE = 3
+    CHANGE_BLOCK_PERIOD = 4
