+++ Src/Services/nomenclature_service.py
@@ -1,0 +1,13 @@
+from Src.Core.abstract_logic import abstract_logic
+from Src.data_reposity import data_reposity
+from Src.Core.validator import operation_exception, validator
+from Src.Models.group import group_model
+from Src.Models.range import range_model
+from Src.Models.nomenclature import nomenclature_model
+
+
+"""
+Сервис для работы с номенклатурой
+"""
+class nomenclature_service(abstract_logic):
+    pass
